Http Cookies work and are done.
Amount of changes in a single branch made are unfortunate, but they were all worth it.

+++ src/IO/TempFileStream.cs
@@ -3,6 +3,7 @@
 namespace Poly.IO {
 
     public class TempFileStream : Stream {
+        public string FileName { get; private set; }
         public FileInfo FileInfo { get; private set; }
         public FileStream BaseStream { get; private set; }
 
@@ -20,16 +21,23 @@ namespace Poly.IO {
         }
 
         public TempFileStream() {
-            FileInfo = new FileInfo(Path.GetTempFileName());
+            FileName = Path.GetTempFileName();
+            FileInfo = new FileInfo(FileName);
             BaseStream = FileInfo.Open(FileMode.Create);
         }
 
         ~TempFileStream() {
             BaseStream.Close();
-            FileInfo.Delete();
+
+            if (FileInfo.FullName == FileName)
+                FileInfo.Delete();
         }
 
-        public override void Flush() => BaseStream.Flush();
+        public virtual void SaveTo(string file_path) =>
+            FileInfo.MoveTo(file_path);
+
+        public override void Flush() => 
+            BaseStream.Flush();
 
         public override int Read(byte[] buffer, int offset, int count) =>
             BaseStream.Read(buffer, offset, count);
